refactor scripts - fix some minor issues
